scraper/pyproject.toml
[project]
name = "google-business-scraper"
version = "0.2.0"
description = "Scrapers for Google Business reviews and opening hours, stored in Strapi"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "TKWS" }]

dependencies = [
    "outscraper",
    "requests",
    "python-dotenv",
    "instaloader",
    "fastapi",
    "uvicorn",
    "browser-cookie3",
]

[project.scripts]
gbr-sync = "google_business_review.main:main"
gbr-schema = "google_business_review.generate_schema:main"
gboh-sync = "google_business_opening_hours.main:main"
gboh-schema = "google_business_opening_hours.generate_schema:main"

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.uv.sources]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.setuptools.packages.find]
where = ["core/src", "google_business_review/src", "google_business_opening_hours/src"]
include = ["core*", "google_business_review*", "google_business_opening_hours*"]

